+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## [1.84.14](https://github.com/adobe/spacecat-api-service/compare/v1.84.13...v1.84.14) (2025-01-10)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v1.61.19 ([#682](https://github.com/adobe/spacecat-api-service/issues/682)) ([5c200b8](https://github.com/adobe/spacecat-api-service/commit/5c200b8b218ebae18fabfafaf811fb938f2511a3))
+
+## [1.84.13](https://github.com/adobe/spacecat-api-service/compare/v1.84.12...v1.84.13) (2025-01-09)
+
+
+### Bug Fixes
+
+* org invalid uuid ([#681](https://github.com/adobe/spacecat-api-service/issues/681)) ([7d9556b](https://github.com/adobe/spacecat-api-service/commit/7d9556b6a127b14c99e50bde4090ee7e39d5bade))
+
+## [1.84.12](https://github.com/adobe/spacecat-api-service/compare/v1.84.11...v1.84.12) (2025-01-09)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#680](https://github.com/adobe/spacecat-api-service/issues/680)) ([49bc749](https://github.com/adobe/spacecat-api-service/commit/49bc7492a003af26b4595b0ba84f958c5d363627))
+
+## [1.84.11](https://github.com/adobe/spacecat-api-service/compare/v1.84.10...v1.84.11) (2025-01-08)
+
+
+### Bug Fixes
+
+* approve site candidate ([#679](https://github.com/adobe/spacecat-api-service/issues/679)) ([f6d77fe](https://github.com/adobe/spacecat-api-service/commit/f6d77fe14ce65bd42540d772e8a8f873be30ecc0))
+
+## [1.84.10](https://github.com/adobe/spacecat-api-service/compare/v1.84.9...v1.84.10) (2025-01-08)
+
+
+### Bug Fixes
+
+* getAllLatestForSite endpoint ([f5fa5ad](https://github.com/adobe/spacecat-api-service/commit/f5fa5ad98fd7d564c2ee54ed93b5de294a6741e2))
+
 ## [1.84.9](https://github.com/adobe/spacecat-api-service/compare/v1.84.8...v1.84.9) (2025-01-07)